+++ tests/dashboard/server.py
@@ -6,8 +6,6 @@ import time
 import livemetrics
 import livemetrics.publishers.http
 
-import tests.publishers
-
 LM = livemetrics.LiveMetrics('{"version":"1.0"}',"Test server",True, True)
 
 LM.gauge("random",lambda: random.randint(1,10))