+++ HISTORY.md
@@ -5,15 +5,16 @@ Release History
 0.3
 ---
 
+- Rework of the dashboard to separate value and display (new format for the description file)
+- Fix memory gauge on recent Linux
+
 0.2 (2019-07-24)
 ----------------
 
 - Evolution of the dashboard
 - Addition of an automatic gauge for cpu and memory (Linux only)
 
-
 0.1 (2019-07-09)
 ----------------
 
 - Initial release
-