+++ HISTORY.md
@@ -6,6 +6,7 @@ Release History
 ---
 
 - Rework of the dashboard to separate value and display (new format for the description file)
+- Make the dashboard configuration file a Jinja2 file (to be able to define variables, macros, etc.)
 - Fix memory gauge on recent Linux
 
 0.2 (2019-07-24)